+++ internal/azure/types/integer_type.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	"fmt"
-	"math"
 
 	"github.com/Azure/terraform-provider-azapi/internal/azure/utils"
 )
@@ -26,16 +25,10 @@ func (t *IntegerType) Validate(body interface{}, path string) []error {
 	}
 	var v int
 	switch input := body.(type) {
-	case float64:
-		if math.Round(input) != input {
-			return []error{utils.ErrorMismatch(path, "integer", fmt.Sprintf("%T", body))}
-		}
-		v = int(input)
-	case float32:
-		if math.Round(float64(input)) != float64(input) {
-			return []error{utils.ErrorMismatch(path, "integer", fmt.Sprintf("%T", body))}
-		}
-		v = int(input)
+	case float64, float32:
+		// TODO: skip validation for now because of the following issue:
+		// the bicep-types-az parses float as integer type and it should be fixed: https://github.com/Azure/bicep-types-az/issues/1404
+		return nil
 	case int64:
 		v = int(input)
 	case int32: