+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
 - Fix a bug when upgrading from previous provider `azapi_resource` resource will set `tags` for resources that don't have `tags` in the configuration.
 - Fix a bug that `azapi_resource` resource cannot handle tags with unknown values.
 - Fix a bug that `null` value can't pass the schema validation.
+- Fix a bug that schema validation fails to validate the float number in the body.
 
 ## v1.13.0
 BREAKING CHANGES: